Refactored the rest into shared functions
