+++ trunk/cli/src/commands/containers.rs
@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use bollard::container::DownloadFromContainerOptions;
+use bollard::container::{CreateContainerOptions, DownloadFromContainerOptions, StartContainerOptions};
 use bollard::exec::{CreateExecOptions, StartExecOptions, StartExecResults};
 use bollard::Docker;
 use elf::endian::AnyEndian;
@@ -8,13 +8,15 @@ use std::fs::File;
 use std::io::Cursor;
 use std::path::Path;
 use bollard::image::BuildImageOptions;
-use bollard::models::BuildInfo;
+use bollard::models::{BuildInfo, HostConfig};
+use bollard::container::{Config};
 
 use crate::commands::generic_build::GenericBuildError;
 use crate::manifest::{Manifest, PackagedFile};
 use crate::sync_utils::{ByteStreamSyncReceiver, ByteStreamSyncSender};
 use futures_util::stream::StreamExt;
 use hyper::Body;
+use rand::Rng;
 use tar::{Archive, Builder, EntryType, Header};
 use tee_readwrite::TeeReader;
 use tokio::task;
@@ -23,15 +25,15 @@ use tokio_task_manager::Task;
 
 /// Used to stop container when dropped, relies on using [tokio_task_manager::TaskManager::wait]
 /// to ensure `Drop` will run to completion
-pub struct ReclaimableContainer<'a> {
-    id: &'a str,
+pub struct ReclaimableContainer {
+    pub id: String,
     docker: Docker,
     task: Task,
 }
 
-impl<'a> ReclaimableContainer<'a> {
+impl ReclaimableContainer {
     #[must_use]
-    pub fn new(name: &'a str, docker: &Docker, task: Task) -> Self {
+    pub fn new(name: String, docker: &Docker, task: Task) -> Self {
         Self {
             id: name,
             docker: docker.clone(),
@@ -40,16 +42,16 @@ impl<'a> ReclaimableContainer<'a> {
     }
 }
 
-impl<'a> Drop for ReclaimableContainer<'a> {
+impl Drop for ReclaimableContainer {
     fn drop(&mut self) {
         let docker = self.docker.clone();
-        let id = self.id.to_string();
+        let id = self.id.clone();
         let handle = tokio::runtime::Handle::current();
         let mut task = self.task.clone();
         handle.spawn(async move {
             println!("Stopping {id}");
             docker
-                .stop_container(&id, None)
+                .stop_container(id.clone().as_str(), None)
                 .await
                 .expect("error stopping container");
             println!("Stopped {id}");
@@ -103,18 +105,133 @@ pub async fn exec_in_container(
     Ok::<String, anyhow::Error>(total_output)
 }
 
+pub async fn run_temporary_container(
+    docker: Docker,
+    image: &str,
+    _task: Task
+) -> Result<ReclaimableContainer, anyhow::Error> {
+
+    let options = Some(CreateContainerOptions {
+        name: image.to_string(),
+        platform: None,
+    });
+
+    let host_config = HostConfig {
+        auto_remove: Some(true),
+        ..Default::default()
+    };
+
+    let config = Config {
+        image: Some(image.to_string()),
+        entrypoint: Some(vec!["sleep".to_string()]),
+        cmd: Some(vec!["300".to_string()]),
+        user: Some("root".to_string()),
+        host_config: Some(host_config),
+        ..Default::default()
+    };
+
+    let container = docker.create_container(options, config).await?;
+    docker
+        .start_container(&container.id, None::<StartContainerOptions<String>>)
+        .await?;
+
+    // This will stop the container, whether we return an error or not
+    return Ok(ReclaimableContainer::new(container.id.clone(), &docker, _task));
+}
+
+
+
+pub async fn find_installed_extension_files(docker: Docker, container_id: &str) -> Result<HashMap<String, Vec<String>>, anyhow::Error> {
+
+    let sharedir = exec_in_container(
+        docker.clone(),
+        container_id,
+        vec!["pg_config", "--sharedir"],
+    )
+        .await?;
+    let sharedir = sharedir.trim();
+
+    let pkglibdir = exec_in_container(
+        docker.clone(),
+        container_id,
+        vec!["pg_config", "--pkglibdir"],
+    )
+        .await?;
+    let pkglibdir = pkglibdir.trim();
+
+    // collect changes from container filesystem
+    println!("Collecting files installed by this extension...");
+    let changes = docker
+        .container_changes(container_id)
+        .await?
+        .expect("Expected to find changed files");
+
+    let mut pkglibdir_list = vec![];
+    let mut sharedir_list = vec![];
+
+    for change in changes {
+        if change.kind == 1
+            && (change.path.ends_with(".so")
+            || change.path.ends_with(".bc")
+            || change.path.ends_with(".sql")
+            || change.path.ends_with(".control"))
+        {
+            if change.path.starts_with(pkglibdir.clone()) {
+                let file_in_pkglibdir = change.path;
+                let file_in_pkglibdir = file_in_pkglibdir.strip_prefix(pkglibdir);
+                let file_in_pkglibdir = file_in_pkglibdir.unwrap();
+                let file_in_pkglibdir = file_in_pkglibdir.trim_start_matches('/');
+                pkglibdir_list.push(file_in_pkglibdir.to_owned());
+            } else if change.path.starts_with(sharedir.clone()) {
+                let file_in_sharedir = change.path;
+                let file_in_sharedir = file_in_sharedir.strip_prefix(sharedir);
+                let file_in_sharedir = file_in_sharedir.unwrap();
+                let file_in_sharedir = file_in_sharedir.trim_start_matches('/');
+                sharedir_list.push(file_in_sharedir.to_owned());
+            } else {
+                println!(
+                    "WARNING: file {} is not in pkglibdir or sharedir",
+                    change.path
+                );
+            }
+        }
+    }
+
+    println!("Sharedir files:");
+    for sharedir_file in sharedir_list.clone() {
+        println!("{sharedir_file}");
+    }
+    println!("Pkglibdir files:");
+    for pkglibdir_file in pkglibdir_list.clone() {
+        println!("{pkglibdir_file}");
+    }
+
+    let mut result = HashMap::new();
+    result.insert("sharedir".to_string(), sharedir_list);
+    result.insert("pkglibdir".to_string(), pkglibdir_list);
+    return Ok(result);
+}
 
 // Build an image
 // The Dockerfile and build directory can be in different directories.
+// The caller provides an image name prefix, and this function returns
+// the complete image name.
 pub async fn build_image(
     docker: Docker,
-    image_name: &str,
+    image_name_prefix: &str,
     dockerfile_path: &str,
     build_directory: &Path,
     build_args: HashMap<&str, &str>
-) -> Result<(), anyhow::Error> {
+) -> Result<String, anyhow::Error> {
 
     let dockerfile = dockerfile_path.to_owned();
+
+    let random_suffix = {
+        let mut rng = rand::thread_rng();
+        rng.gen_range(0..1000000).to_string()
+    };
+
+    let image_name = format!("{}{}", image_name_prefix.to_owned(), &random_suffix);
 
     let (receiver, sender, stream) = ByteStreamSyncSender::new();
 
@@ -180,25 +297,43 @@ pub async fn build_image(
             }
         }
     }
-
-    return Ok(());
+    return Ok(image_name.to_string());
 }
 
 // Scan sharedir and package lib dir from a Trunk builder container for files from a provided list.
 // Package these files into a Trunk package.
-pub async fn copy_from_container_into_package(
+pub async fn package_installed_extension_files(
     docker: Docker,
     container_id: &str,
     package_path: &str,
-    sharedir: &str,
-    pkglibdir: &str,
-    sharedir_list: Vec<String>,
-    pkglibdir_list: Vec<String>,
     extension_name: &str,
     extension_version: &str,
 ) -> Result<(), anyhow::Error> {
+
     let extension_name = extension_name.to_owned();
     let extension_version = extension_version.to_owned();
+
+    let sharedir = exec_in_container(
+        docker.clone(),
+        container_id,
+        vec!["pg_config", "--sharedir"],
+    )
+        .await?;
+    let sharedir = sharedir.trim();
+
+    let pkglibdir = exec_in_container(
+        docker.clone(),
+        container_id,
+        vec!["pg_config", "--pkglibdir"],
+    )
+        .await?;
+    let pkglibdir = pkglibdir.trim();
+
+    let extension_files = find_installed_extension_files(docker.clone(), container_id).await?;
+
+    let sharedir_list = extension_files["sharedir"].clone();
+    let pkglibdir_list = extension_files["pkglibdir"].clone();
+
     let pkglibdir = pkglibdir.to_owned();
     let sharedir = sharedir.to_owned();
